refactor: 削除インタラクションをDeleteObjectServiceへ分離
DeleteObjectStateはモード遷移とライフサイクル(OnEnter/Update委譲/OnExit)のみ担い、
ホバー/ドラッグ選択/一括削除/ESCキャンセル/プレビュー片付けはDeleteObjectService.Update()・
Cleanup()へ移動。挙動は不変、200行制約も緩和。
