+++ moorestech_client/Assets/Scripts/Client.Game/InGame/UI/UIState/State/DeleteObjectState.cs
@@ -1,16 +1,11 @@
-using System;
-using Client.Game.InGame.Block;
 using Client.Game.InGame.Control;
-using Client.Game.InGame.Entity.Object;
 using Client.Game.InGame.Train.RailGraph;
 using Client.Game.InGame.UI.KeyControl;
-using Client.Game.InGame.UI.Tooltip;
 using Client.Game.InGame.UI.UIState.Input;
 using Client.Game.InGame.UI.UIState.State.DragDelete;
 using Client.Game.InGame.UI.UIState.UIObject;
 using Client.Input;
 using UnityEngine;
-using UnityEngine.EventSystems;
 
 namespace Client.Game.InGame.UI.UIState.State
 {
@@ -20,11 +15,7 @@ namespace Client.Game.InGame.UI.UIState.State
 
         private readonly ScreenClickableCameraController _screenClickableCameraController;
 
-        private IDeleteTarget _deleteTargetObject;
-        private bool _isRemoveDeniedReasonShown;
-
-        private readonly DragDeleteSelection _selection = new();
-        private bool _isDragging;
+        private readonly DeleteObjectService _deleteObjectService = new();
 
         private readonly RailGraphClientCache _railGraphClientCache;
 
@@ -45,39 +36,14 @@ namespace Client.Game.InGame.UI.UIState.State
 
         public UITransitContext GetNextUpdate()
         {
-            // 拒否理由ツールチップを毎フレーム先に消す
-            // Reset the denial-reason tooltip at the start of each frame
-            if (_isRemoveDeniedReasonShown)
-            {
-                MouseCursorTooltip.Instance.Hide();
-                _isRemoveDeniedReasonShown = false;
-            }
-
-            // モード遷移を判定する（ESCはモードを抜けない）
-            // Handle mode transitions (ESC no longer exits the mode)
+            // モード遷移を判定する（ESCはモードを抜けず削除サービス側で選択キャンセルに使う）
+            // Handle mode transitions (ESC stays in the mode and is used as selection cancel by the delete service)
             var transit = HandleTransition();
             if (transit != null) return transit;
 
-            // ESCは選択キャンセルとして扱いモードに留まる
-            // Treat ESC as a selection cancel while staying in this mode
-            if (InputManager.UI.CloseUI.GetKeyDown) CancelOnEscape();
-
-            // カーソル下の削除対象を取得（無ければnull）
-            // Resolve the target hovered this frame (null when nothing hit)
-            BlockClickDetectUtil.TryGetCursorOnComponent(out IDeleteTarget hovered);
-
-            // 左クリック開始でドラッグ選択を開始する
-            // Begin a drag selection on left-click down
-            HandleDragStart();
-
-            // ドラッグ中は選択へ追加、非ドラッグは単体表示
-            // While dragging accumulate the selection, otherwise show single hover preview
-            if (_isDragging) UpdateDragSelection();
-            else UpdateSingleHoverPreview();
-
-            // 左クリック離しで選択を確定して削除する
-            // Commit and delete the selection on left-click release
-            HandleRelease();
+            // 削除インタラクションはサービスに委譲する
+            // Delegate the delete interaction to the service
+            _deleteObjectService.Update();
 
             _screenClickableCameraController.GetNextUpdate();
             return null;
@@ -94,104 +60,13 @@ namespace Client.Game.InGame.UI.UIState.State
                 return null;
             }
 
-            void CancelOnEscape()
-            {
-                _selection.CancelSelection();
-
-                // 単体ホバー中の赤プレビューも確実に戻す
-                // Also clear any single-hover red preview to be safe
-                if (!_isDragging && _deleteTargetObject != null)
-                {
-                    _deleteTargetObject.ResetMaterial();
-                    _deleteTargetObject = null;
-                }
-            }
-
-            void HandleDragStart()
-            {
-                if (!InputManager.Playable.ScreenLeftClick.GetKeyDown) return;
-                if (EventSystem.current.IsPointerOverGameObject()) return;
-
-                // 単体プレビューの所有権を選択側へ渡す
-                // Hand off preview ownership from single-hover to the selection
-                if (_deleteTargetObject != null)
-                {
-                    _deleteTargetObject.ResetMaterial();
-                    _deleteTargetObject = null;
-                }
-
-                _selection.BeginDrag();
-                _isDragging = true;
-            }
-
-            void UpdateDragSelection()
-            {
-                // キャンセル済みドラッグは何もしない（ESC後は離すまで不活性）
-                // A canceled drag is inert until the button is released
-                if (!_selection.CanCommit()) return;
-
-                if (hovered == null) return;
-
-                if (hovered.IsRemovable(out var reason))
-                {
-                    _selection.AddTarget(hovered);
-                }
-                else
-                {
-                    MouseCursorTooltip.Instance.Show(reason, isLocalize: false);
-                    _isRemoveDeniedReasonShown = true;
-                }
-            }
-
-            void UpdateSingleHoverPreview()
-            {
-                // ホバー対象が変われば旧を戻し新を表示する
-                // Swap preview when the hovered target changes
-                if (hovered != null)
-                {
-                    if (_deleteTargetObject == null || _deleteTargetObject != hovered)
-                    {
-                        if (_deleteTargetObject != null) _deleteTargetObject.ResetMaterial();
-                        _deleteTargetObject = hovered;
-                        _deleteTargetObject.SetRemovePreviewing();
-                    }
-                }
-                else if (_deleteTargetObject != null)
-                {
-                    _deleteTargetObject.ResetMaterial();
-                    _deleteTargetObject = null;
-                }
-
-                // 削除不可な対象は理由ツールチップだけ表示する
-                // For a non-removable target only show the denial tooltip
-                if (_deleteTargetObject != null && !_deleteTargetObject.IsRemovable(out var reason))
-                {
-                    MouseCursorTooltip.Instance.Show(reason, isLocalize: false);
-                    _isRemoveDeniedReasonShown = true;
-                }
-            }
-
-            void HandleRelease()
-            {
-                if (!InputManager.Playable.ScreenLeftClick.GetKeyUp) return;
-
-                if (_isDragging && _selection.CanCommit()) _selection.CommitDelete();
-                _isDragging = false;
-            }
-
             #endregion
         }
 
-
         public void OnExit()
         {
-            if (_deleteTargetObject != null) _deleteTargetObject.ResetMaterial();
+            _deleteObjectService.Cleanup();
             _deleteBarObject.gameObject.SetActive(false);
-
-            // モード離脱時に進行中の選択プレビューを片付ける
-            // Clear any in-progress selection previews when leaving the mode
-            _selection.CancelSelection();
-            _isDragging = false;
 
             _screenClickableCameraController.OnExit();
         }
