+++ moorestech_client/Assets/Scripts/Client.Game/InGame/UI/UIState/State/DragDelete/DeleteObjectService.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 287a698c8d104badae9789976027c283
+timeCreated: 1781483382
